refactor(train_bert): split into dataset_bert.py, checkpoint_bert.py, and slimmed train_bert.py (Closes #36)
